fix for when no files are in bucket

+++ import.py
@@ -45,55 +45,57 @@ while True:
     # iterate through files in bucket
     s3 = boto3.resource('s3')
     files = s3.meta.client.list_objects_v2(Bucket=config_dict['s3_bucket'])
-    filenames = [file['Key'] for file in files['Contents']]
+    # if there are no files, then contents key will not exist
+    if 'Contents' in files:
+        filenames = [file['Key'] for file in files['Contents']]
 
-    for filename in filenames:
-        try:
-            # check if the file is valid
-            is_csv = filename.split('.')[-1] == 'csv'
-            has_site_id = filename.split('_')[0].isdigit()
+        for filename in filenames:
+            try:
+                # check if the file is valid
+                is_csv = filename.split('.')[-1] == 'csv'
+                has_site_id = filename.split('_')[0].isdigit()
 
-            if is_csv and has_site_id:
-                logging.info("Importing file {}".format(filename))
+                if is_csv and has_site_id:
+                    logging.info("Importing file {}".format(filename))
 
-                # remove the folder path from filename
-                filename_local = filename.split('/')[-1]
+                    # remove the folder path from filename
+                    filename_local = filename.split('/')[-1]
 
-                # download csv from S3 and open with a csv dict reader
-                s3.meta.client.download_file(config_dict['s3_bucket'], filename, filename_local)
-                csvfile = open(filename_local)
-                reader = csv.DictReader(csvfile)
+                    # download csv from S3 and open with a csv dict reader
+                    s3.meta.client.download_file(config_dict['s3_bucket'], filename, filename_local)
+                    csvfile = open(filename_local)
+                    reader = csv.DictReader(csvfile)
 
-                # get site id from folder name prefix
-                site_id = int(filename.split('_')[0])
+                    # get site id from folder name prefix
+                    site_id = int(filename.split('_')[0])
 
-                # load ids of existing entries in db. select all rows from table, filter by site id, then select only SeqNos
-                from sqlalchemy.sql import select
-                s = select([alarms_table.c.SeqNo]).where(alarms_table.c.site_id == site_id)
-                result = conn.execute(s)
-                rows = result.fetchall()
-                # put into a set for more efficient lookups
-                seqnos = set([row[0] for row in rows])
+                    # load ids of existing entries in db. select all rows from table, filter by site id, then select only SeqNos
+                    from sqlalchemy.sql import select
+                    s = select([alarms_table.c.SeqNo]).where(alarms_table.c.site_id == site_id)
+                    result = conn.execute(s)
+                    rows = result.fetchall()
+                    # put into a set for more efficient lookups
+                    seqnos = set([row[0] for row in rows])
 
-                # build list of rows to add. only include a row if it doesn't already exist in db
-                to_add = []
-                for row in reader:
-                    if not int(row['SeqNo']) in seqnos:
-                        row['site_id'] = site_id
-                        to_add.append(row)
+                    # build list of rows to add. only include a row if it doesn't already exist in db
+                    to_add = []
+                    for row in reader:
+                        if not int(row['SeqNo']) in seqnos:
+                            row['site_id'] = site_id
+                            to_add.append(row)
 
-                # insert to db
-                if len(to_add) > 0:
-                    conn.execute(alarms_table.insert(), to_add)
+                    # insert to db
+                    if len(to_add) > 0:
+                        conn.execute(alarms_table.insert(), to_add)
 
-                # delete file locally
-                os.remove(filename_local)
+                    # delete file locally
+                    os.remove(filename_local)
 
-                # delete file on s3 bucket
-                s3.meta.client.delete_object(Bucket=config_dict['s3_bucket'], Key=filename)
+                    # delete file on s3 bucket
+                    s3.meta.client.delete_object(Bucket=config_dict['s3_bucket'], Key=filename)
 
-        except Exception as e:
-            logging.error("Exception {} when trying to import {}".format(e, filename))
+            except Exception as e:
+                logging.error("Exception {} when trying to import {}".format(e, filename))
 
     # sleep
     time.sleep(config_dict['loop_seconds'])
